helpers/password.py
from datetime import timedelta

import jwt
import bcrypt

from configs import JWT_ALGORITHM, JWT_SECRET_KEY


def hash_password(password: str) -> str:
    salt = bcrypt.gensalt()
    hashed_password = bcrypt.hashpw(password.encode("utf-8"), salt)
    return hashed_password.decode("utf-8")


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return bcrypt.checkpw(
        plain_password.encode("utf-8"), hashed_password.encode("utf-8")
    )


def create_access_token(data: dict, expire_at: timedelta = None):
    to_encode = data.copy()
    to_encode.update({"exp": expire_at})
    return jwt.encode(to_encode, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
