pyproject.toml
[tool.poetry]
name = "keystroke-recognition-api"
version = "0.1.0"
description = "Keystroke Dynamics Recognition"
authors = ["Khoa Tran <tdkhoa0303@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.0"
fastapi = {extras = ["standard"], version = "^0.115.3"}
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
supabase = "^2.9.1"
celery = {extras = ["redis"], version = "^5.4.0"}
motor = "^3.6.0"
pyjwt = "^2.9.0"
pandas = "^2.2.3"
xgboost = "^2.1.2"
scikit-learn = "^1.5.2"
imblearn = "^0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false